% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjRsq.R
\name{adjRsq}
\alias{adjRsq}
\title{Function to factorize dataset loaded from HDFS}
\usage{
adjRsq(testDataset, predictionDataset, nObs, nPred, isLossLog)
}
\arguments{
\item{List}{returned by function 'readDataFromHDFS()'}
}
\value{
List: [[1]] Sparklyr connection object, [[2]] Factorized training dataset [[3]] Factorized testing dataset
}
\description{
Function to factorize dataset loaded from HDFS
}

