% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapFactorizer.R
\name{mapFactorizer}
\alias{mapFactorizer}
\title{Function to factorize dataset loaded from HDFS}
\usage{
mapFactorizer(datasetToFactorize)
}
\arguments{
\item{List}{returned by function 'readDataFromHDFS()'}
}
\value{
List: [[1]] Sparklyr connection object, [[2]] Factorized training dataset [[3]] Factorized testing dataset
}
\description{
Function to factorize dataset loaded from HDFS
}

