% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageInstaller.R
\name{packageInstaller}
\alias{packageInstaller}
\title{Function to factorize dataset loaded from HDFS}
\usage{
packageInstaller()
}
\arguments{
\item{List}{returned by function 'readDataFromHDFS()'}
}
\value{
List: [[1]] Sparklyr connection object, [[2]] Factorized training dataset [[3]] Factorized testing dataset
}
\description{
Function to factorize dataset loaded from HDFS
}

