% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readInitialDataFromHDFS.R
\name{readInitialDataFromHDFS}
\alias{readInitialDataFromHDFS}
\title{Function loads data from my HDFS}
\usage{
readInitialDataFromHDFS(sc)
}
\arguments{
\item{none, }{names are fixed}
}
\value{
list of 3 elements: spark connection reference, train dataset in spark ref., test dataset in spark ref.,
}
\description{
Function loads data from my HDFS
}

